transactions/views.py
from rest_framework import viewsets
from rest_framework import permissions
from .models import Transaction
from .serializer import TransactionSerializer


class TransactionViewSet(viewsets.ModelViewSet):
    """
    API endpoint that allows users to be viewed or edited.
    """
    queryset = Transaction.objects.all().order_by('id')
    serializer_class = TransactionSerializer
    permission_classes = [permissions.AllowAny]
