requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
django==4.1.7
djangorestframework==3.14.0
psycopg2==2.9.5
python-dotenv==1.0.0
pytz==2022.7.1
sqlparse==0.4.3
